requirements.txt
Django==1.8
Pillow==2.9.0
argparse==1.2.1
django-admin-tools==0.6.0
django-appconf==1.0.1
django-autoslug==1.8.0
django-compressor==1.5
django-seo-js==0.3.1
django-solo==1.1.0
django-wysiwyg-redactor==0.4.8.1
djangorestframework==3.1.3
psycopg2==2.6.1
pytils==0.3
requests==2.2.1
six==1.9.0
wsgiref==0.1.2
